pyproject.toml
[project]
name = "fitness-agent"
version = "0.1.0"
description = "Fitness and meal planning agent with local + cloud LLM routing"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "pydantic-ai[anthropic,openai]>=1.5.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "greenlet>=3.0.0",
    "alembic>=1.14.0",
    "apscheduler>=3.10.4",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.12",
    "structlog>=24.4.0",
    "pypdf>=5.0.0",
    "docx2txt>=0.8",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "psycopg2-binary>=2.9.10",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
